+++ src/main/java/org/bytesync/hotelmanagement/repository/UserRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     @Query("""
     select u from User u
@@ -54,5 +54,5 @@ public interface UserRepository extends JpaRepository<User, Long> {
     ) from User u
     where u.id = :id
 """)
-    Optional<UserDetailsDto> findDetailsById(Long id);
+    Optional<UserDetailsDto> findDetailsById(Integer id);
 }
